refactor: extract partial eval tests on its own file
